Decouple autoplanner motorpool projection
Keep reserve projection eligibility in the theater domain so commander code does not depend on mission-generation internals.
