+++ game/commander/tasks/primitive/motorpool.py
@@ -30,9 +30,7 @@ class PlanMotorpoolAttack(PackagePlanningTask[MotorpoolGroundObject]):
         self.propose_common_escorts()
 
     def _rendered_unit_count(self) -> int:
-        from game.missiongenerator.motorpoolpopulator import (
-            motorpool_rendered_unit_count,
-        )
+        from game.theater.theatergroundobject import motorpool_rendered_unit_count
 
         settings = self.target.coalition.game.settings
         return motorpool_rendered_unit_count(
