+++ game/missiongenerator/motorpoolpopulator.py
@@ -3,9 +3,11 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from game.dcs.groundunittype import GroundUnitType
-from game.ground_forces.ai_ground_planner import reserve_armor_for
 from game.theater.theatergroup import TheaterGroup, TheaterUnit
-from game.theater.theatergroundobject import MotorpoolGroundObject
+from game.theater.theatergroundobject import (
+    MotorpoolGroundObject,
+    motorpool_projected_counts,
+)
 from game.point_with_heading import PointWithHeading
 
 if TYPE_CHECKING:
@@ -17,67 +19,6 @@ _COLUMNS = 5
 # Keep the Garage_A building at the authored marker; start vehicles clear of it
 # behind the building. 150 ft is the authoring-friendly value.
 _GRID_OFFSET_M = 45.72
-
-
-def _select_capped(
-    reserve: dict[GroundUnitType, int], cap: int
-) -> dict[GroundUnitType, int]:
-    """Proportionally reduce ``reserve`` so its counts sum to at most ``cap``,
-    using the largest-remainder method (keeps a representative spread of types).
-    Returns a copy of ``reserve`` unchanged when it already fits under the cap."""
-    total = sum(reserve.values())
-    if total <= cap:
-        return {ut: n for ut, n in reserve.items() if n > 0}
-    exact = {ut: count * cap / total for ut, count in reserve.items()}
-    floors = {ut: int(v) for ut, v in exact.items()}
-    remaining = cap - sum(floors.values())
-    if remaining > 0:
-        by_frac = sorted(
-            ((ut, exact[ut] - floors[ut]) for ut in reserve),
-            key=lambda kv: kv[1],
-            reverse=True,
-        )
-        for ut, _frac in by_frac[:remaining]:
-            floors[ut] += 1
-    return {ut: n for ut, n in floors.items() if n > 0}
-
-
-def _projected_counts(
-    motorpools: list[MotorpoolGroundObject], cap: int
-) -> list[dict[GroundUnitType, int]]:
-    reserve = reserve_armor_for(motorpools[0].control_point)
-    selected = _select_capped(reserve, cap)
-    per_tgo: list[dict[GroundUnitType, int]] = [{} for _ in motorpools]
-    slot = 0
-    for unit_type, count in selected.items():
-        for _ in range(count):
-            bucket = per_tgo[slot % len(motorpools)]
-            bucket[unit_type] = bucket.get(unit_type, 0) + 1
-            slot += 1
-    return per_tgo
-
-
-def motorpool_rendered_unit_count(
-    tgo: MotorpoolGroundObject, motorpool_enabled: bool, spawn_cap: int
-) -> int:
-    """Return the alive units in the current or next motorpool snapshot."""
-    if not motorpool_enabled or spawn_cap <= 0:
-        return 0
-    motorpools = [
-        candidate
-        for candidate in tgo.control_point.ground_objects
-        if isinstance(candidate, MotorpoolGroundObject)
-    ]
-    if tgo not in motorpools:
-        return 0
-    projected_count = sum(
-        _projected_counts(motorpools, spawn_cap)[motorpools.index(tgo)].values()
-    )
-    if tgo.groups:
-        # Groups are ephemeral and may outlive a reserve decrement from the prior
-        # mission. Never plan more units than the next current snapshot can render.
-        return min(tgo.alive_unit_count, projected_count)
-    return projected_count
 
 
 class MotorpoolPopulator:
@@ -110,7 +51,7 @@ class MotorpoolPopulator:
         # each TGO with the full reserve independently would render — and on a
         # strike decrement — the same reserve unit once per TGO, corrupting
         # base.armor when a CP has more than one authored motorpool location.
-        for tgo, counts in zip(motorpools, _projected_counts(motorpools, cap)):
+        for tgo, counts in zip(motorpools, motorpool_projected_counts(motorpools, cap)):
             self._build_groups(tgo, counts)
 
     def _build_groups(
