+++ game/theater/theatergroundobject.py
@@ -8,6 +8,7 @@ from typing import Any, Iterator, List, Optional, TYPE_CHECKING
 from dcs.mapping import Point
 from shapely.geometry import Point as ShapelyPoint
 
+from game.ground_forces.ai_ground_planner import reserve_armor_for
 from game.sidc import (
     Entity,
     LandEquipmentEntity,
@@ -31,6 +32,68 @@ if TYPE_CHECKING:
     from game.threatzones import ThreatPoly
     from .theatergroup import TheaterUnit, TheaterGroup
     from .controlpoint import ControlPoint, Coalition
+
+
+def _select_capped(
+    reserve: dict[GroundUnitType, int], cap: int
+) -> dict[GroundUnitType, int]:
+    """Proportionally reduce ``reserve`` to at most ``cap`` units."""
+    total = sum(reserve.values())
+    if total <= cap:
+        return {ut: n for ut, n in reserve.items() if n > 0}
+    exact = {ut: count * cap / total for ut, count in reserve.items()}
+    floors = {ut: int(v) for ut, v in exact.items()}
+    remaining = cap - sum(floors.values())
+    if remaining > 0:
+        by_frac = sorted(
+            ((ut, exact[ut] - floors[ut]) for ut in reserve),
+            key=lambda kv: kv[1],
+            reverse=True,
+        )
+        for ut, _frac in by_frac[:remaining]:
+            floors[ut] += 1
+    return {ut: n for ut, n in floors.items() if n > 0}
+
+
+def motorpool_projected_counts(
+    motorpools: list[MotorpoolGroundObject], cap: int
+) -> list[dict[GroundUnitType, int]]:
+    """Project one shared reserve pool across a CP's motorpool locations."""
+    reserve = reserve_armor_for(motorpools[0].control_point)
+    selected = _select_capped(reserve, cap)
+    per_tgo: list[dict[GroundUnitType, int]] = [{} for _ in motorpools]
+    slot = 0
+    for unit_type, count in selected.items():
+        for _ in range(count):
+            bucket = per_tgo[slot % len(motorpools)]
+            bucket[unit_type] = bucket.get(unit_type, 0) + 1
+            slot += 1
+    return per_tgo
+
+
+def motorpool_rendered_unit_count(
+    tgo: MotorpoolGroundObject, motorpool_enabled: bool, spawn_cap: int
+) -> int:
+    """Return the alive units in the current or next motorpool snapshot."""
+    if not motorpool_enabled or spawn_cap <= 0:
+        return 0
+    motorpools = [
+        candidate
+        for candidate in tgo.control_point.ground_objects
+        if isinstance(candidate, MotorpoolGroundObject)
+    ]
+    if tgo not in motorpools:
+        return 0
+    projected_count = sum(
+        motorpool_projected_counts(motorpools, spawn_cap)[
+            motorpools.index(tgo)
+        ].values()
+    )
+    if tgo.groups:
+        # Groups are ephemeral and may outlive a reserve decrement from the prior
+        # mission. Never plan more units than the next current snapshot can render.
+        return min(tgo.alive_unit_count, projected_count)
+    return projected_count
 
 
 NAME_BY_CATEGORY = {
